Refactored application name from lms_app to universities_app
